+++ CHANGES.rst
@@ -4,7 +4,9 @@ CHANGES
 1.0.1-1 (unreleased)
 ====================
 
-- Nothing changed yet.
+- Render d3_cloud's script tag with conditional comment for IE >=9 (what
+  ``js.d3`` supports).
+  [davidjb]
 
 
 1.0.1 (2012-12-21)